+++ simulation/behavior_tree_plugin/src/action_node.cpp
@@ -47,7 +47,7 @@ void ActionNode::getBlackBoardValues()
   if (!getInput<std::shared_ptr<hdmap_utils::HdMapUtils>>("hdmap_utils", hdmap_utils)) {
     THROW_SIMULATION_ERROR("failed to get input hdmap_utils in ActionNode");
   }
-  if (!getInput<std::shared_ptr<traffic_simulator::TrafficLightManager>>(
+  if (!getInput<std::shared_ptr<traffic_simulator::TrafficLightManagerBase>>(
         "traffic_light_manager", traffic_light_manager)) {
     THROW_SIMULATION_ERROR("failed to get input traffic_light_manager in ActionNode");
   }