+++ simulation/simple_sensor_simulator/include/simple_sensor_simulator/sensor_simulation/detection_sensor/detection_sensor.hpp
@@ -17,6 +17,7 @@
 
 #include <simulation_api_schema.pb.h>
 
+#include <autoware_auto_perception_msgs/msg/predicted_objects.hpp>
 #include <autoware_perception_msgs/msg/dynamic_object_array.hpp>
 #include <memory>
 #include <rclcpp/rclcpp.hpp>
@@ -25,24 +26,56 @@
 
 namespace simple_sensor_simulator
 {
-class DetectionSensor
+class DetectionSensorBase
 {
+protected:
+  double last_update_stamp_;
+
+  simulation_api_schema::DetectionSensorConfiguration configuration_;
+
+  explicit DetectionSensorBase(
+    const double last_update_stamp,
+    const simulation_api_schema::DetectionSensorConfiguration & configuration)
+  : last_update_stamp_(last_update_stamp), configuration_(configuration)
+  {
+  }
+
 public:
-  DetectionSensor(
+  virtual void update(
+    const double, const std::vector<traffic_simulator_msgs::EntityStatus> &, const rclcpp::Time &,
+    const std::vector<std::string> &) = 0;
+};
+
+template <typename T>
+class DetectionSensor : public DetectionSensorBase
+{
+  const typename rclcpp::Publisher<T>::SharedPtr publisher_ptr_;
+
+public:
+  explicit DetectionSensor(
     const double current_time,
     const simulation_api_schema::DetectionSensorConfiguration & configuration,
-    std::shared_ptr<rclcpp::Publisher<autoware_perception_msgs::msg::DynamicObjectArray>>
-      publisher_ptr);
-  void update(
-    double current_time, const std::vector<traffic_simulator_msgs::EntityStatus> & status,
-    const rclcpp::Time & stamp, const std::vector<std::string> & detected_objects);
+    const typename rclcpp::Publisher<T>::SharedPtr & publisher)
+  : DetectionSensorBase(current_time, configuration), publisher_ptr_(publisher)
+  {
+  }
 
-private:
-  simulation_api_schema::DetectionSensorConfiguration configuration_;
-  std::shared_ptr<rclcpp::Publisher<autoware_perception_msgs::msg::DynamicObjectArray>>
-    publisher_ptr_;
-  double last_update_stamp_;
+  auto update(
+    const double, const std::vector<traffic_simulator_msgs::EntityStatus> &, const rclcpp::Time &,
+    const std::vector<std::string> &) -> void override;
 };
+
+template <>
+void DetectionSensor<autoware_perception_msgs::msg::DynamicObjectArray>::update(
+  const double, const std::vector<traffic_simulator_msgs::EntityStatus> &, const rclcpp::Time &,
+  const std::vector<std::string> &);
+
+#ifndef SCENARIO_SIMULATOR_V2_BACKWARD_COMPATIBLE_TO_AWF_AUTO
+template <>
+void DetectionSensor<autoware_auto_perception_msgs::msg::PredictedObjects>::update(
+  const double, const std::vector<traffic_simulator_msgs::EntityStatus> &, const rclcpp::Time &,
+  const std::vector<std::string> &);
+#endif
 }  // namespace simple_sensor_simulator
 
 #endif  // SIMPLE_SENSOR_SIMULATOR__SENSOR_SIMULATION__DETECTION_SENSOR__DETECTION_SENSOR_HPP_