+++ simulation/simulation_interface/include/simulation_interface/conversions.hpp
@@ -24,6 +24,10 @@
 #include <std_msgs.pb.h>
 #include <traffic_simulator_msgs.pb.h>
 
+#ifndef SCENARIO_SIMULATOR_V2_BACKWARD_COMPATIBLE_TO_AWF_AUTO
+#include <autoware_auto_perception_msgs/msg/traffic_signal.hpp>
+#endif
+
 #include <autoware_control_msgs/msg/control_command.hpp>
 #include <autoware_perception_msgs/msg/traffic_light_state.hpp>
 #include <autoware_vehicle_msgs/msg/shift.hpp>
@@ -85,14 +89,6 @@ void toMsg(const traffic_simulator_msgs::Axle & proto, traffic_simulator_msgs::m
 void toProto(
   const traffic_simulator_msgs::msg::Axles & axles, traffic_simulator_msgs::Axles & proto);
 void toMsg(const traffic_simulator_msgs::Axles & proto, traffic_simulator_msgs::msg::Axles & axles);
-/*
-void toProto(
-  const traffic_simulator_msgs::msg::Property & p,
-  traffic_simulator_msgs::Property & proto);
-void toMsg(
-  const traffic_simulator_msgs::Property & proto,
-  traffic_simulator_msgs::msg::Property & p);
-*/
 void toProto(
   const traffic_simulator_msgs::msg::VehicleParameters & p,
   traffic_simulator_msgs::VehicleParameters & proto);
@@ -160,6 +156,11 @@ void toMsg(
 void toProto(
   const autoware_perception_msgs::msg::TrafficLightState & traffic_light_state,
   simulation_api_schema::TrafficLightState & proto);
+#ifndef SCENARIO_SIMULATOR_V2_BACKWARD_COMPATIBLE_TO_AWF_AUTO
+void toProto(
+  const autoware_auto_perception_msgs::msg::TrafficSignal & traffic_light_state,
+  simulation_api_schema::TrafficLightState & proto);
+#endif
 }  // namespace simulation_interface
 
 #endif  // SIMULATION_INTERFACE__CONVERSIONS_HPP_