+++ simulation/traffic_simulator/src/behavior/route_planner.cpp
@@ -95,6 +95,16 @@ void RoutePlanner::cancelGoal(const traffic_simulator_msgs::msg::LaneletPose & e
   }
 }
 
+std::vector<geometry_msgs::msg::Pose> RoutePlanner::getGoalPosesInWorldFrame()
+{
+  const auto lanelet_poses = getGoalPoses();
+  std::vector<geometry_msgs::msg::Pose> ret;
+  for (const auto & lanelet_pose : lanelet_poses) {
+    ret.emplace_back(hdmap_utils_ptr_->toMapPose(lanelet_pose).pose);
+  }
+  return ret;
+}
+
 std::vector<traffic_simulator_msgs::msg::LaneletPose> RoutePlanner::getGoalPoses()
 {
   std::vector<traffic_simulator_msgs::msg::LaneletPose> goal_poses;